+++ R/check_method_params.R
@@ -23,7 +23,7 @@ check_method_params <- function(y, ite, params) {
   if (length(ratio_dis) == 0) {
     ratio_dis <- 0.5
   } else {
-    if (!inherits(ratio_dis, "numeric") | (ratio_dis < 0) | (ratio_dis > 1)) {
+    if (!inherits(ratio_dis, "numeric") || (ratio_dis < 0) || (ratio_dis > 1)) {
       stop("Invalid 'ratio_dis' input. Please input a number between 0 and 1.")
     }
   }
@@ -35,12 +35,12 @@ check_method_params <- function(y, ite, params) {
   if (length(ite_method_dis) == 0) {
     ite_method_dis <- "aipw"
   } else {
-    if (!(ite_method_dis %in% c("aipw", "slearner","tlearner","xlearner",
-                                "bart","bcf","cf","tpoisson"))) {
+    if (!(ite_method_dis %in% c("aipw", "slearner", "tlearner", "xlearner",
+                                "bart", "cf", "tpoisson"))) {
       stop(paste(
-        "Invalid ITE method for Discovery Subsample. Please choose ",
+        "Invalid ITE method for discovery subsample. Please choose ",
         "from the following:\n", "'aipw', 'bart', 'slearner','tlearner', ",
-        "'xlearner', 'bcf', 'cf', or 'tpoisson'"
+        "'xlearner', 'cf', or 'tpoisson'"
       ))
     }
   }
@@ -50,12 +50,12 @@ check_method_params <- function(y, ite, params) {
   if (length(ite_method_inf) == 0) {
     ite_method_inf <- "aipw"
   } else {
-    if (!(ite_method_dis %in% c("aipw", "slearner","tlearner","xlearner",
-                                "bart","bcf", "cf", "tpoisson"))) {
+    if (!(ite_method_inf %in% c("aipw", "slearner", "tlearner", "xlearner",
+                                "bart", "cf", "tpoisson"))) {
       stop(paste(
-        "Invalid ITE method for Inference Subsample. Please choose ",
+        "Invalid ITE method for inference subsample. Please choose ",
         "from the following:\n", "'aipw', 'bart', 'slearner','tlearner', ",
-        "'xlearner', 'bcf', 'cf', or 'tpoisson'"
+        "'xlearner', 'cf', or 'tpoisson'"
       ))
     }
   }
@@ -65,7 +65,8 @@ check_method_params <- function(y, ite, params) {
   # Propensity Score Estimation Parameters Check--------------------------------
 
   ps_method_dis <- getElement(params, "ps_method_dis")
-  if (!(ite_method_dis %in% c("slearner", "tlearner", "xlearner", "tpoisson"))) {
+  if (!(ite_method_dis %in% c("slearner", "tlearner",
+                              "xlearner", "tpoisson"))) {
     if (length(ps_method_dis) == 0) {
       ps_method_dis <- "SL.xgboost"
     } else {
@@ -80,7 +81,8 @@ check_method_params <- function(y, ite, params) {
   params[["ps_method_dis"]] <- ps_method_dis
 
   ps_method_inf <- getElement(params, "ps_method_inf")
-  if (!(ite_method_inf %in% c("slearner", "tlearner", "xlearner", "tpoisson"))) {
+  if (!(ite_method_inf %in% c("slearner", "tlearner",
+                              "xlearner", "tpoisson"))) {
     if (length(ps_method_inf) == 0) {
       ps_method_inf <- "SL.xgboost"
     } else {