+++ R/cre.R
@@ -7,7 +7,8 @@
 #'
 #' @param y An observed response vector.
 #' @param z A treatment vector.
-#' @param X A covariate matrix (or a data frame).
+#' @param X A covariate matrix (or a data frame). Should be provided as
+#' numerical values.
 #' @param method_params The list of parameters to define the models used,
 #' including:
 #'   - *Parameters for Honest Splitting*
@@ -18,21 +19,22 @@
 #'     (default: 'aipw').
 #'     - *ps_method_dis*: The estimation model for the propensity score on the
 #'       discovery subsample (default: 'SL.xgboost').
-#'     - *or_method_dis*: The estimation model for the outcome regressions
+#'     - *oreg_method_dis*: The estimation model for the outcome regressions
 #'       estimate_ite_aipw on the discovery subsample (default: 'SL.xgboost').
 #'   - *Parameters for Inference*
 #'     - *ite_method_inf*: The method to estimate the inference sample ITE
 #'     (default: 'aipw').
 #'     - *ps_method_inf*: The estimation model for the propensity score on the
 #'       inference subsample (default: 'SL.xgboost').
-#'     - *or_method_inf*: The estimation model for the outcome regressions in
+#'     - *oreg_method_inf*: The estimation model for the outcome regressions in
 #'       estimate_ite_aipw on the inference subsample (default: 'SL.xgboost').
-#' @param hyper_params The list of hyper parameters to finetune the method,
+#' @param hyper_params The list of hyper parameters to fine-tune the method,
 #' including:
-#'  - *intervention_vars*: Intervention-able variables used for Rules Generation
-#'  (default: NULL).
+#'  - *intervention_vars*: Intervention-able variables used for rules
+#'  generation. Use `NULL` to include all variables (default: `NULL`).
 #'  - *offset*: Name of the covariate to use as offset (i.e. 'x1') for
-#'     T-Poisson ITE Estimation. NULL if offset is not used (default: NULL).
+#'     T-Poisson ITE estimation. Use `NULL` if offset is not used
+#'     (default: NULL).
 #'  - *ntrees_rf*: A number of decision trees for random forest (default: 20).
 #'  - *ntrees_gbm*: A number of decision trees for the generalized boosted
 #' regression modeling algorithm.
@@ -42,13 +44,14 @@
 #'  - *max_depth*: Maximum rules length (default: 3).
 #'  - *replace*: Boolean variable for replacement in bootstrapping for
 #'  rules generation by random forest (default: TRUE).
-#'  - *t_decay*: The decay threshold for rules pruning (default: 0.025).
-#'  - *t_ext*: The threshold to define too generic or too specific (extreme)
-#'  rules (default: 0.01, range: (0,0.5)).
+#'  - *t_decay*: The decay threshold for rules pruning. Higher values will
+#'  carry out an aggressive pruning (default: 0.025).
+#'  - *t_ext*: The threshold to truncate too generic or too specific (extreme)
+#'  rules (default: 0.01, range: [0, 0.5)).
 #'  - *t_corr*: The threshold to define correlated rules (default: 1,
 #'  range: (0,+inf)).
 #'  - *t_pvalue*: the threshold to define statistically significant rules
-#' (default: 0.05, range: (0,1)).
+#' (default: 0.05, range: (0, 1)).
 #'  - *stability_selection*: Whether or not using stability selection for
 #'  selecting the rules (default: TRUE).
 #'  - *cutoff*:  Threshold (percentage) defining the minimum cutoff value for
@@ -61,15 +64,19 @@
 #' @param ite The estimated ITE vector. If given both the ITE estimation steps
 #' in Discovery and Inference are skipped (default: NULL).
 #'
+#'
 #' @return
 #' An S3 object containing:
 #' - A number of Decision Rules extracted at each step (`M`).
 #' - A data.frame of Conditional Average Treatment Effect decomposition
 #' estimates with corresponding uncertainty quantification (`CATE`).
-#' - A list of Method Parameters (`method_params`).
-#' - A list of Hyper Parameters (`hyper_params`).
+#' - A list of method parameters (`method_params`).
+#' - A list of hyper parameters (`hyper_params`).
 #' - An Individual Treatment Effect predicted (`ite_pred`).
 #'
+#' @note
+#' - If `intervention_vars` are provided, it's important to note that the
+#' individual treatment effect will still be computed using all covariates.
 #' @export
 #'
 #' @examples
@@ -120,6 +127,7 @@ cre <- function(y, z, X,
   st_time_cre <- proc.time()
 
   # Input checks ---------------------------------------------------------------
+  check_input_data(y, z, X, ite)
   method_params <- check_method_params(y = y,
                                        ite = ite,
                                        params = method_params)
@@ -127,7 +135,6 @@ cre <- function(y, z, X,
                                      params = hyper_params)
 
   # Honest Splitting -----------------------------------------------------------
-  X_names <- names(as.data.frame(X))
   subgroups <- honest_splitting(y, z, X,
                                 getElement(method_params, "ratio_dis"), ite)
   discovery <- subgroups[["discovery"]]
@@ -143,6 +150,7 @@ cre <- function(y, z, X,
   X_inf <- inference$X
   ite_inf <- inference$ite
 
+  intervention_vars <- getElement(hyper_params, "intervention_vars")
 
   # Discovery ------------------------------------------------------------------
   logger::log_info("Starting rules discovery...")
@@ -150,15 +158,20 @@ cre <- function(y, z, X,
   # Estimate ITE
   if (is.null(ite)) {
     ite_dis <- estimate_ite(y = y_dis, z = z_dis, X = X_dis,
-                      ite_method = getElement(method_params, "ite_method_dis"),
-                      ps_method = getElement(method_params, "ps_method_dis"),
-                      oreg_method = getElement(method_params,"oreg_method_dis"),
-                      offset = getElement(method_params, "offset"))
+                    ite_method = getElement(method_params, "ite_method_dis"),
+                    ps_method = getElement(method_params, "ps_method_dis"),
+                    oreg_method = getElement(method_params, "oreg_method_dis"),
+                    offset = getElement(method_params, "offset"))
   } else {
     logger::log_info("Using the provided ITE estimations...")
   }
 
-  # Generate Decision Rules
+  # Filter only Intervention-able variables
+  if (!is.null(intervention_vars)) {
+    X_dis <- X_dis[, intervention_vars, drop = FALSE]
+  }
+
+  # Discover Decision Rules
   discovery <- discover_rules(X_dis,
                               ite_dis,
                               method_params,
@@ -176,22 +189,27 @@ cre <- function(y, z, X,
   # Estimate ITE
   if (is.null(ite)) {
     ite_inf <- estimate_ite(y = y_inf, z = z_inf, X = X_inf,
-                      ite_method = getElement(method_params, "ite_method_inf"),
-                      ps_method = getElement(method_params, "ps_method_inf"),
-                      oreg_method = getElement(method_params, "oreg_method_inf"),
-                      offset = getElement(method_params,"offset"))
+                    ite_method = getElement(method_params, "ite_method_inf"),
+                    ps_method = getElement(method_params, "ps_method_inf"),
+                    oreg_method = getElement(method_params, "oreg_method_inf"),
+                    offset = getElement(method_params, "offset"))
   } else {
     logger::log_info("Skipped generating ITE.",
                      "The provided ITE will be used.")
   }
 
+  # Filter only Intervention-able variables
+  if (!is.null(intervention_vars)) {
+    X_inf <- X_inf[, intervention_vars, drop = FALSE]
+  }
+
   # Generate rules matrix
   if (length(rules) == 0) {
     rules_matrix_inf <- NA
-    rules_explicit <- NA
+    rules_explicit <- c()
   } else {
     rules_matrix_inf <- generate_rules_matrix(X_inf, rules)
-    rules_explicit <- interpret_rules(rules, X_names)
+    rules_explicit <- interpret_rules(rules, colnames(X_inf))
   }
 
   # Estimate CATE
@@ -201,12 +219,14 @@ cre <- function(y, z, X,
 
   # Estimate ITE
   if (M["select_significant"] > 0) {
+    # Filter only Intervention-able variables
+    if (!is.null(intervention_vars)) X <- X[, intervention_vars, drop = FALSE]
     rules_matrix <- generate_rules_matrix(X, rules)
     filter <- rules_explicit %in%
               cate_inf$summary$Rule[2:length(cate_inf$summary$Rule)]
     rules_df <- as.data.frame(rules_matrix[, filter])
     names(rules_df) <- rules_explicit[filter]
-    ite_pred <- predict(cate_inf$model, rules_df)
+    ite_pred <- predict(cate_inf$model, rules_df) + cate_inf$summary$Estimate[1]
   } else {
     ite_pred <- cate_inf$summary$Estimate[1]
   }
@@ -229,7 +249,7 @@ cre <- function(y, z, X,
   # Return Results -------------------------------------------------------------
   end_time_cre <- proc.time()
   logger::log_info("Done with running CRE function!",
-                   "(WC: {g_wc_str(st_time_cre, end_time_cre)}",".)")
+                   "(WC: {g_wc_str(st_time_cre, end_time_cre)}", ".)")
   logger::log_info("Done!")
   return(results)
 }