+++ R/generate_cre_dataset.R
@@ -54,7 +54,7 @@ generate_cre_dataset <- function(n = 1000, rho = 0, n_rules = 2, p = 10,
   if (!(binary_outcome %in% c(TRUE, FALSE))) {
     stop("Invalid 'binary' input. Please specify TRUE or FALSE.")
   }
-  if (is.numeric(n) & !is.integer(n)) {
+  if (is.numeric(n) && !is.integer(n)) {
     n <- as.integer(n)
   }
 
@@ -85,7 +85,7 @@ generate_cre_dataset <- function(n = 1000, rho = 0, n_rules = 2, p = 10,
     if (confounding == "lin") {
       mean <- X$x1 + X$x3 + X$x4
     } else if (confounding == "nonlin") {
-      mean <- X$x1 + cos(X$x3)
+      mean <- X$x1 + cos(X$x3*X$x4)
     } else if (confounding == "no") {
       mean <- 0
     } else {
@@ -109,11 +109,11 @@ generate_cre_dataset <- function(n = 1000, rho = 0, n_rules = 2, p = 10,
                  "rules has not been implemented yet. ",
                  "Available 'n_rules' options: {1,2}."))
     }
-    y0[X$x4 > 0.5] <- y0[X$x4 > 0.5] + 2*effect_size
+    y0[X$x4 > 0.5] <- y0[X$x4 > 0.5] + 0.5 * effect_size
   }
   if (n_rules >= 4) {
     y1[X$x5 <= 0.5 & X$x7 > 0.5 & X$x8 <= 0.5] <-
-                y1[X$x5 <= 0.5 & X$x7 > 0.5 & X$x8 <= 0.5] + 2*effect_size
+                y1[X$x5 <= 0.5 & X$x7 > 0.5 & X$x8 <= 0.5] + 2 * effect_size
   }
   if (n_rules >= 5) {
     stop(paste("Synthtic dataset with continuos outcome and ", n_rules,