+++ functional_tests/discovery.R
@@ -1,39 +1,75 @@
 set.seed(2022)
-library(foreach)
-library(doParallel)
+library("foreach")
+library("doParallel")
+library("CRE")
 
 # Set Experiment Parameter
-n_rules <- 2
-sample_size <- 2000
-effect_sizes <- seq(0, 4, 0.2)
-confoundings <- c("no","lin","nonlin")
-ITE_estimators <- c("aipw","cf","bcf","slearner","tlearner","xlearner","bart")
-n_seeds <- 100
-ratio_dis <- 0.5
+experiment <- "main"
+cutoff <- 0.8
 
-# Set Ground Truth
-{
-  if (n_rules==1) {
-    dr <- c("x1>0.5 & x2<=0.5")
-    em <- c("x1","x2")
-  } else if (n_rules==2) {
-    dr <- c("x1>0.5 & x2<=0.5", "x5>0.5 & x6<=0.5")
-    em <- c("x1","x2","x5","x6")
-  } else if (n_rules==3) {
-    dr <- c("x1>0.5 & x2<=0.5", "x5>0.5 & x6<=0.5", "x4>0.5")
-    em <- c("x1","x2","x5","x6","x4")
-  } else if (n_rules==4) {
-    dr <- c("x1>0.5 & x2<=0.5", "x5>0.5 & x6<=0.5",
-            "x4>0.5", "x5<=0.5 & x7>0.5 & x8<=0.5")
-    em <- c("x1","x2","x5","x6","x4","x7","x8")
-  } else {
-    stop(paste("Synthtic dataset with", n_rules,"rules has not been",
-               "implemented yet. Available 'n_rules' options: {1,2,3,4}."))
-  }
+if (experiment=="main") {
+  n_rules <- 2
+  sample_size <- 2000
+  confounding <- "lin"
+  dr <- c("x1>0.5 & x2<=0.5", "x5>0.5 & x6<=0.5")
+  em <- c("x1","x2","x5","x6")
+  max_depth <- 2
+} else if (experiment=="small_sample") {
+  n_rules <- 2
+  sample_size <- 1000
+  confounding <- "lin"
+  dr <- c("x1>0.5 & x2<=0.5", "x5>0.5 & x6<=0.5")
+  em <- c("x1","x2","x5","x6")
+  max_depth <- 2
+} else if (experiment=="big_sample") {
+  n_rules <- 2
+  sample_size <- 5000
+  confounding <- "lin"
+  dr <- c("x1>0.5 & x2<=0.5", "x5>0.5 & x6<=0.5")
+  em <- c("x1","x2","x5","x6")
+  max_depth <- 2
+} else if (experiment=="more_rules") {
+  n_rules <- 4
+  sample_size <- 2000
+  confounding <- "lin"
+  dr <- c("x1>0.5 & x2<=0.5", "x5>0.5 & x6<=0.5",
+          "x4>0.5", "x5<=0.5 & x7>0.5 & x8<=0.5")
+  em <- c("x1","x2","x5","x6","x4","x7","x8")
+  max_depth <- 3
+} else if (experiment=="rct") {
+  n_rules <- 2
+  sample_size <- 2000
+  confounding <- "no"
+  dr <- c("x1>0.5 & x2<=0.5", "x5>0.5 & x6<=0.5")
+  em <- c("x1","x2","x5","x6")
+  max_depth <- 2
+} else if (experiment=="nonlin_conf") {
+  n_rules <- 2
+  sample_size <- 2000
+  confounding <- "nonlin"
+  dr <- c("x1>0.5 & x2<=0.5", "x5>0.5 & x6<=0.5")
+  em <- c("x1","x2","x5","x6")
+  max_depth <- 2
+} else if (experiment=="personalize") {
+  n_rules <- NA
+  sample_size <- NA
+  confounding <- NA
+  dr <- NA
+  em <- NA
+  max_depth <- NA
+  stop("'personalize' esperiment non defined yet.")
+} else {
+  stop(paste("'",experiment,"' experiment doesn't exist.", sep=""))
 }
 
-# Set Method and Hyper Parameters
+
+# Other Setting
 {
+  n_seeds <- 250
+  ratio_dis <- 0.5
+  effect_sizes <- seq(0, 5, 0.2)
+  ITE_estimators <- c("aipw","cf",'slearner","tlearner","xlearner","bart")
+
   method_params <- list(ratio_dis = ratio_dis,
                         ite_method_dis = "aipw",
                         ps_method_dis = "SL.xgboost",
@@ -47,86 +83,59 @@ ratio_dis <- 0.5
                        ntrees_rf = 40,
                        ntrees_gbm = 40,
                        node_size = 20,
-                       max_nodes = 8,
-                       max_depth = 3,
+                       max_nodes = 2^max_depth,
+                       max_depth = max_depth,
                        t_decay = 0.025,
                        t_ext = 0.01,
                        t_corr = 1,
                        t_pvalue = 0.05,
                        replace = TRUE,
                        stability_selection = TRUE,
-                       cutoff = 0.8,
+                       cutoff = cutoff,
                        pfer = 1,
                        penalty_rl = 1)
+
+  evaluate <- function(ground_truth, prediction) {
+    intersect <- intersect(prediction, ground_truth)
+    union <- union(prediction, ground_truth)
+    TP <- length(intersect)
+    FP <- length(setdiff(prediction, ground_truth))
+    FN <- length(setdiff(ground_truth, prediction))
+    recall <- TP / (TP + FN) # quantity
+    precision <- TP / (TP + FP) # quality
+    IoU <- length(intersect) / length(union)
+    evaluate <- list(recall = recall,
+                     precision = precision,
+                     IoU = IoU)
+    return(evaluate)
+  }
+  extract_effect_modifiers <- function(rules_list, X_names) {
+    effect_modifiers <- c()
+    for (X_name in X_names) {
+      if (any(grepl(X_name, rules_list))) {
+        effect_modifiers <- append(effect_modifiers, X_name)
+      }
+    }
+    return(effect_modifiers)
+  }
 }
 
+
 # Set Cluster
-cl <- makeCluster(detectCores())
+cl <- makeCluster(detectCores(), type="PSOCK")
+clusterExport(cl, ls(globalenv()))
+clusterEvalQ(cl, library("CRE"))
 registerDoParallel(cl)
 
-# Run Experiments (confounding x effect_size x method x seed)
-for (confounding in confoundings) {
-  print(paste("Confounding:",confounding))
-  discovery <- data.frame(matrix(ncol = 9, nrow = 0))
-  for(effect_size in effect_sizes){
-    print(paste("Effect Size", effect_size))
-
-    # CRE
-    for (ITE_estimator in ITE_estimators){
-      # CRE (estimator i)
-      method <- paste("CRE (", ITE_estimator, ")", sep = "")
-      time.before <- proc.time()
-      discovery_i <- foreach(seed = seq(1, n_seeds, 1), .combine=rbind) %dopar% {
-        library(devtools)
-        load_all()
-        set.seed(seed)
-
-        # Generate Dataset
-        dataset <- generate_cre_dataset(n = sample_size,
-                                        rho = 0,
-                                        p = 10,
-                                        effect_size = effect_size,
-                                        n_rules = n_rules,
-                                        binary_covariates = TRUE,
-                                        binary_outcome = FALSE,
-                                        confounding = confounding)
-        y <- dataset[["y"]]
-        z <- dataset[["z"]]
-        X <- dataset[["X"]]
-        X_names <- colnames(X)
-
-        method_params[["ite_method_dis"]] <- ITE_estimator
-        method_params[["ite_method_inf"]] <- ITE_estimator
-        hyper_params[["pfer"]] <- n_rules/(effect_size+1)
-        metrics <- tryCatch({
-          result <- cre(y, z, X, method_params, hyper_params)
-
-          dr_pred <- result$CATE$Rule[result$CATE$Rule %in% "(BATE)" == FALSE]
-          metrics_dr <- evaluate(dr, dr_pred)
-          em_pred <- extract_effect_modifiers(dr_pred, X_names)
-          metrics_em <- evaluate(em, em_pred)
-
-          c(method, effect_size, seed,
-            metrics_dr$IoU, metrics_dr$precision, metrics_dr$recall,
-            metrics_em$IoU, metrics_em$precision, metrics_em$recall)
-          },
-          error = function(e) {
-            c(method, effect_size, seed, NaN,NaN,NaN,NaN,NaN,NaN)
-          })
-        return(metrics)
-      }
-      discovery <- rbind(discovery, discovery_i)
-      time.after <- proc.time()
-      print(paste("CRE -", ITE_estimator,"(Time: ",
-                  round((time.after - time.before)[[3]],2), "sec)"))
-    }
-
-    # HCT
+discovery <- data.frame(matrix(ncol = 9, nrow = 0))
+for(effect_size in effect_sizes){
+  print(paste("Effect Size", effect_size))
+  for (ITE_estimator in ITE_estimators){
+    # CRE (estimator i)
+    method <- paste("CRE (", ITE_estimator, ")", sep = "")
     time.before <- proc.time()
     discovery_i <- foreach(seed = seq(1, n_seeds, 1), .combine=rbind) %dopar% {
-      library(devtools)
-      library(causalTree)
-      load_all()
+      library("CRE")
       set.seed(seed)
 
       # Generate Dataset
@@ -143,88 +152,45 @@ for (confounding in confoundings) {
       X <- dataset[["X"]]
       X_names <- colnames(X)
 
-      subgroups <- honest_splitting(y, z, X, ratio_dis)
-      discovery <- subgroups[["discovery"]]
-      inference <- subgroups[["inference"]]
-      y_dis <- discovery$y
-      z_dis <- discovery$z
-      X_dis <- discovery$X
-      y_inf <- inference$y
-      z_inf <- inference$z
-      X_inf <- inference$X
-      data_dis <- as.data.frame(cbind(y_dis, X_dis))
-      data_inf <- as.data.frame(cbind(y_inf, z_inf, X_inf))
+      method_params[["ite_method_dis"]] <- ITE_estimator
+      method_params[["ite_method_inf"]] <- ITE_estimator
+      hyper_params[["pfer"]] <- n_rules/(effect_size+1)
+      metrics <- tryCatch({
+        result <- cre(y, z, X, method_params, hyper_params)
 
-      fit.tree <- causalTree(y_dis ~ ., data = data_dis, treatment = z_dis,
-                             split.Rule = "CT", cv.option = "CT",
-                             split.Honest = T, cv.Honest = T, maxdepth = 4)
-      opt.cp <- fit.tree$cptable[,1][which.min(fit.tree$cptable[,4])]
-      pruned <- prune(fit.tree, opt.cp)
+        dr_pred <- result$CATE$Rule[result$CATE$Rule %in% "(BATE)" == FALSE]
+        metrics_dr <- evaluate(dr, dr_pred)
+        em_pred <- extract_effect_modifiers(dr_pred, X_names)
+        metrics_em <- evaluate(em, em_pred)
 
-      # Extract Causal Decision Rules
-      rules <- as.numeric(row.names(pruned$frame[pruned$numresp]))
-      rules.ctree <- vector("list",length(rules))
-      for (k in rules[-1]){
-        sub <- as.data.frame(matrix(NA,
-                                    nrow = 1,
-                                    ncol = nrow(as.data.frame(
-                                               path.rpart(pruned,
-                                                          node=k,
-                                                          print.it = FALSE)))-1)
-                                                )
-        capture.output(for (h in 1:ncol(sub)){
-          sub[,h] <- as.character(print(as.data.frame(
-                                            path.rpart(pruned,
-                                                       node=k,
-                                                       print.it=FALSE))[h+1,1]))
-          sub_pop <- noquote(paste(sub , collapse = " & "))
-        })
-        subset <- with(data_inf, data_inf[which(eval(parse(text=sub_pop))),])
-        if (length(unique(subset$z_inf))!= 1){
-          rules.ctree[[k]] <- sub_pop
-        }
-      }
-      leaves_dfs <- rep(FALSE,length(rules))
-      leaves_dfs[unique(pruned$where)] <- TRUE
-      dr_pred <- unlist(rules.ctree[rules[leaves_dfs]])
-      for (i in 1:length(dr_pred)){
-        dr_pred[i] <- gsub("< ", "<=", dr_pred[i])
-        dr_pred[i] <- gsub(">=", ">", dr_pred[i])
-      }
-      # Predict
-      ite_pred <- predict(pruned, X)
-
-      metrics_dr <- evaluate(dr,dr_pred)
-
-      em_pred <- extract_effect_modifiers(dr_pred, X_names)
-      metrics_em <- evaluate(em,em_pred)
-
-      return(c("HCT", effect_size, seed,
-               metrics_dr$IoU,
-               metrics_dr$precision,
-               metrics_dr$recall,
-               metrics_em$IoU,
-               metrics_em$precision,
-               metrics_em$recall))
+        c(method, effect_size, seed,
+          metrics_dr$IoU, metrics_dr$precision, metrics_dr$recall,
+          metrics_em$IoU, metrics_em$precision, metrics_em$recall)
+      },
+      error = function(e) {
+        c(method, effect_size, seed, NaN,NaN,NaN,NaN,NaN,NaN)
+      })
+      return(metrics)
     }
-    discovery <- rbind(discovery,discovery_i)
+    discovery <- rbind(discovery, discovery_i)
     time.after <- proc.time()
-    print(paste("HCT (Time: ",round((time.after - time.before)[[3]],2), "sec)"))
+    print(paste("CRE -", ITE_estimator,"(Time: ",
+                round((time.after - time.before)[[3]],2), "sec)"))
   }
-  colnames(discovery) <- c("method","effect_size","seed",
-                           "dr_IoU","dr_Precision","dr_Recall",
-                           "em_IoU","em_Precision","em_Recall")
-  rownames(discovery) <- 1:nrow(discovery)
-
-  # Save results
-  results_dir <- "../functional_tests/results/"
-  if (!dir.exists(results_dir)) {
-    dir.create(results_dir)
-  }
-  exp_name <- paste(sample_size,"s_",n_rules,"r_",confounding, sep="")
-  file_dir <- paste(results_dir,"discovery_",exp_name,".RData", sep="")
-  save(discovery,
-       file=file_dir)
 }
+
+colnames(discovery) <- c("method","effect_size","seed",
+                         "dr_IoU","dr_Precision","dr_Recall",
+                         "em_IoU","em_Precision","em_Recall")
+rownames(discovery) <- 1:nrow(discovery)
+
+# Save results
+results_dir <- "~/CRE/functional_tests/results/"
+if (!dir.exists(results_dir)) {
+  dir.create(results_dir)
+}
+exp_name <- paste("discovery",experiment,"cutoff",cutoff, sep="_")
+file_dir <- paste(results_dir,exp_name,".RData", sep="")
+save(discovery, file=file_dir)
 
 stopCluster(cl)